Fix bug where book always displays read
when displaying the status of book it checks if the status is true or false but when book.status is retrieved the value is "false" a string so it will ways be evaluated as true. By converting it to a boolean value and saving it as boolean fixes the bug.